@@ -15,96 +15,67 @@
     "verdaccio": "./bin/verdaccio"
   },
   "dependencies": {
-    "@verdaccio/file-locking": "0.0.7",
-    "@verdaccio/local-storage": "1.2.0",
-    "@verdaccio/streams": "1.0.0",
+    "@verdaccio/local-storage": "2.0.0-beta.3",
+    "@verdaccio/streams": "2.0.0-beta.0",
     "@verdaccio/ui-theme": "1.0.1",
-    "JSONStream": "1.3.4",
-    "async": "2.6.1",
+    "JSONStream": "1.3.5",
+    "async": "3.0.1-0",
     "body-parser": "1.18.3",
     "bunyan": "1.8.12",
-    "chalk": "2.4.1",
-    "commander": "2.18.0",
+    "chalk": "2.4.2",
+    "commander": "2.19.0",
     "compression": "1.7.3",
-    "cookies": "0.7.2",
-    "cors": "2.8.4",
-    "date-fns": "1.29.0",
-    "express": "4.16.3",
+    "cookies": "0.7.3",
+    "cors": "2.8.5",
+    "date-fns": "1.30.1",
+    "express": "4.16.4",
     "global": "4.3.2",
-    "handlebars": "4.0.12",
-    "http-errors": "1.7.1",
-    "js-base64": "2.4.9",
+    "handlebars": "4.1.0",
+    "http-errors": "1.7.2",
+    "js-base64": "2.5.1",
     "js-string-escape": "1.0.1",
-    "js-yaml": "3.12.0",
-    "jsonwebtoken": "8.3.0",
+    "js-yaml": "3.12.2",
+    "jsonwebtoken": "8.5.0",
     "lockfile": "1.0.4",
     "lodash": "4.17.11",
-    "lunr-mutable-indexes": "2.3.1",
-    "marked": "0.5.1",
-    "mime": "2.3.1",
+    "lunr-mutable-indexes": "2.3.2",
+    "marked": "0.6.1",
+    "mime": "2.4.0",
     "minimatch": "3.0.4",
     "mkdirp": "0.5.1",
     "mv": "2.1.1",
     "pkginfo": "0.4.1",
     "request": "2.88.0",
-    "semver": "5.5.1",
+    "semver": "5.6.0",
     "verdaccio-audit": "1.1.0",
-    "verdaccio-htpasswd": "1.0.1",
+    "verdaccio-htpasswd": "2.0.0-beta.1",
     "verror": "1.10.0"
   },
   "devDependencies": {
-    "@babel/cli": "7.2.3",
-    "@babel/core": "7.2.2",
-    "@babel/node": "7.2.2",
-    "@babel/plugin-proposal-class-properties": "7.2.3",
-    "@babel/plugin-proposal-decorators": "7.2.3",
-    "@babel/plugin-proposal-export-namespace-from": "7.2.0",
-    "@babel/plugin-proposal-function-sent": "7.2.0",
-    "@babel/plugin-proposal-json-strings": "7.2.0",
-    "@babel/plugin-proposal-numeric-separator": "7.2.0",
-    "@babel/plugin-proposal-object-rest-spread": "7.2.0",
-    "@babel/plugin-proposal-throw-expressions": "7.2.0",
-    "@babel/plugin-syntax-dynamic-import": "7.2.0",
-    "@babel/plugin-syntax-import-meta": "7.2.0",
-    "@babel/plugin-transform-async-to-generator": "7.2.0",
-    "@babel/plugin-transform-classes": "7.2.2",
-    "@babel/plugin-transform-runtime": "7.2.0",
-    "@babel/polyfill": "7.2.3",
-    "@babel/preset-env": "7.2.3",
-    "@babel/preset-flow": "7.0.0",
-    "@babel/preset-react": "7.0.0",
-    "@babel/register": "7.0.0",
-    "@babel/runtime": "^7.2.0",
-    "@commitlint/cli": "7.2.1",
-    "@commitlint/config-conventional": "7.1.2",
-    "@material-ui/core": "3.1.0",
-    "@material-ui/icons": "3.0.1",
-    "@verdaccio/types": "4.1.4",
+    "@commitlint/cli": "7.5.2",
+    "@commitlint/config-conventional": "7.5.0",
+    "@material-ui/core": "3.9.0",
+    "@material-ui/icons": "3.0.2",
+    "@verdaccio/babel-preset": "0.0.4",
+    "@verdaccio/types": "5.0.0-beta.3",
     "autosuggest-highlight": "3.1.1",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "10.0.1",
-    "babel-jest": "23.6.0",
-    "babel-loader": "8.0.4",
-    "babel-plugin-dynamic-import-node": "2.2.0",
-    "babel-plugin-emotion": "9.2.10",
-    "babel-plugin-flow-runtime": "0.17.0",
-    "bundlesize": "0.17.0",
-    "codecov": "3.1.0",
+    "bundlesize": "0.17.1",
+    "codecov": "3.2.0",
     "cross-env": "5.2.0",
     "css-loader": "0.28.10",
     "emotion": "9.2.12",
-    "enzyme": "3.6.0",
-    "enzyme-adapter-react-16": "1.5.0",
-    "eslint": "5.10.0",
-    "eslint-config-google": "0.11.0",
-    "eslint-config-prettier": "3.3.0",
-    "eslint-loader": "2.1.1",
+    "enzyme": "3.9.0",
+    "enzyme-adapter-react-16": "1.10.0",
+    "eslint": "5.14.1",
+    "eslint-config-google": "0.12.0",
+    "eslint-config-prettier": "4.1.0",
+    "eslint-loader": "2.1.2",
     "eslint-plugin-babel": "5.3.0",
-    "eslint-plugin-flowtype": "3.2.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-jest": "22.1.2",
-    "eslint-plugin-jsx-a11y": "6.1.2",
-    "eslint-plugin-prettier": "3.0.0",
+    "eslint-plugin-flowtype": "3.4.2",
+    "eslint-plugin-import": "2.16.0",
+    "eslint-plugin-jest": "22.3.0",
+    "eslint-plugin-jsx-a11y": "6.2.1",
+    "eslint-plugin-prettier": "3.0.1",
     "eslint-plugin-react": "7.11.1",
     "eslint-plugin-verdaccio": "0.0.5",
     "file-loader": "2.0.0",
@@ -116,52 +87,53 @@
     "husky": "0.15.0-rc.8",
     "identity-obj-proxy": "3.0.0",
     "in-publish": "2.0.0",
-    "jest": "23.6.0",
-    "jest-environment-jsdom": "23.4.0",
-    "jest-environment-jsdom-global": "1.1.0",
-    "jest-environment-node": "23.4.0",
+    "jest": "24.1.0",
+    "jest-environment-jsdom": "24.0.0",
+    "jest-environment-jsdom-global": "1.1.1",
+    "jest-environment-node": "24.0.0",
     "lint-staged": "7.3.0",
-    "localstorage-memory": "1.0.2",
-    "mini-css-extract-plugin": "0.4.3",
-    "node-mocks-http": "1.7.0",
-    "node-sass": "4.9.3",
-    "normalize.css": "8.0.0",
+    "localstorage-memory": "1.0.3",
+    "mini-css-extract-plugin": "0.5.0",
+    "node-mocks-http": "1.7.3",
+    "node-sass": "4.11.0",
+    "normalize.css": "8.0.1",
     "optimize-css-assets-webpack-plugin": "5.0.1",
     "ora": "1.4.0",
     "prettier": "1.14.3",
-    "prop-types": "15.6.2",
+    "prop-types": "15.7.2",
     "puppeteer": "1.8.0",
-    "react": "16.4.2",
+    "react": "16.8.3",
     "react-autosuggest": "9.4.2",
-    "react-dom": "16.4.2",
+    "react-dom": "16.8.3",
     "react-emotion": "9.2.12",
-    "react-hot-loader": "4.2.0",
+    "react-hot-loader": "4.7.1",
     "react-router": "4.3.1",
-    "react-router-dom": "4.2.2",
-    "resolve-url-loader": "3.0.0",
-    "rimraf": "2.6.2",
+    "react-router-dom": "4.3.1",
+    "resolve-url-loader": "3.0.1",
+    "rimraf": "2.6.3",
     "sass-loader": "7.1.0",
     "source-map-loader": "0.2.4",
     "standard-version": "4.4.0",
-    "style-loader": "0.23.0",
-    "stylelint": "9.9.0",
+    "style-loader": "0.23.1",
+    "stylelint": "9.10.1",
     "stylelint-config-recommended": "2.1.0",
     "stylelint-config-recommended-scss": "3.2.0",
     "stylelint-config-styled-components": "0.1.1",
-    "stylelint-processor-styled-components": "1.5.1",
-    "stylelint-scss": "3.3.1",
+    "stylelint-processor-styled-components": "1.5.2",
+    "stylelint-scss": "3.5.4",
     "stylelint-webpack-plugin": "0.10.5",
-    "supertest": "3.3.0",
+    "supertest": "3.4.2",
     "typeface-roboto": "0.0.54",
-    "url-loader": "1.1.1",
+    "url-loader": "1.1.2",
     "verdaccio-auth-memory": "0.0.4",
-    "verdaccio-memory": "1.0.3",
+    "verdaccio-memory": "2.0.0-beta.0",
     "webpack": "4.20.2",
-    "webpack-bundle-analyzer": "3.0.2",
-    "webpack-cli": "3.1.1",
-    "webpack-dev-server": "3.1.14",
-    "webpack-merge": "4.1.4",
-    "whatwg-fetch": "3.0.0"
+    "webpack-bundle-analyzer": "3.0.4",
+    "webpack-cli": "3.2.3",
+    "webpack-dev-server": "3.2.1",
+    "webpack-merge": "4.2.1",
+    "whatwg-fetch": "3.0.0",
+    "xss": "1.0.3"
   },
   "keywords": [
     "private",
