@@ -1,51 +1,68 @@
+/**
+ * @prettier
+ */
+
 import React, { Component, Fragment } from 'react';
 import isNil from 'lodash/isNil';
 
 import storage from './utils/storage';
 import { makeLogin, isTokenExpire } from './utils/login';
 
-import Footer from './components/Footer';
 import Loading from './components/Loading';
 import LoginModal from './components/Login';
 import Header from './components/Header';
 import { Container, Content } from './components/Layout';
-import Route from './router';
+import RouterApp from './router';
 import API from './utils/api';
 import './styles/typeface-roboto.scss';
 import './styles/main.scss';
 import 'normalize.css';
+import Footer from './components/Footer';
+
+export const AppContext = React.createContext();
+export const AppContextProvider = AppContext.Provider;
+export const AppContextConsumer = AppContext.Consumer;
 
 export default class App extends Component {
   state = {
     error: {},
-    logoUrl: '',
+    logoUrl: window.VERDACCIO_LOGO,
     user: {},
-    scope: (window.VERDACCIO_SCOPE) ? `${window.VERDACCIO_SCOPE}:` : '',
+    scope: window.VERDACCIO_SCOPE ? `${window.VERDACCIO_SCOPE}:` : '',
     showLoginModal: false,
     isUserLoggedIn: false,
     packages: [],
     isLoading: true,
-  }
+  };
 
   componentDidMount() {
-    this.loadLogo();
     this.isUserAlreadyLoggedIn();
-    this.loadPackages();
+    this.loadOnHandler();
   }
 
   // eslint-disable-next-line no-unused-vars
   componentDidUpdate(_, prevState) {
     const { isUserLoggedIn } = this.state;
     if (prevState.isUserLoggedIn !== isUserLoggedIn) {
-      this.loadPackages();
+      this.loadOnHandler();
     }
   }
 
-    loadLogo = () => {
-    const logoUrl = window.VERDACCIO_LOGO;
-    this.setState({
-      logoUrl,
-    });
+  render() {
+    const { isLoading, isUserLoggedIn, packages, logoUrl, user, scope } = this.state;
+
+    return (
+      <Container isLoading={isLoading}>
+        {isLoading ? (
+          <Loading />
+        ) : (
+          <Fragment>
+            <AppContextProvider value={{ isUserLoggedIn, packages, logoUrl, user, scope }}>{this.renderContent()}</AppContextProvider>
+          </Fragment>
+        )}
+        {this.renderLoginModal()}
+      </Container>
+    );
   }
 
   isUserAlreadyLoggedIn = () => {
@@ -53,16 +70,16 @@ export default class App extends Component {
     const token = storage.getItem('token');
     const username = storage.getItem('username');
     if (isTokenExpire(token) || isNil(username)) {
-     this.handleLogout();
+      this.handleLogout();
     } else {
       this.setState({
         user: { username, token },
         isUserLoggedIn: true,
       });
     }
-  }
+  };
 
-  loadPackages = async () => {
+  loadOnHandler = async () => {
     try {
       this.req = await API.request('packages', 'GET');
       this.setState({
@@ -70,40 +87,37 @@ export default class App extends Component {
         isLoading: false,
       });
     } catch (error) {
-      this.handleShowAlertDialog({
+      // FIXME: add dialog
+      console.error({
         title: 'Warning',
         message: `Unable to load package list: ${error.message}`,
       });
       this.setLoading(false);
     }
-  }
+  };
 
-  setLoading = isLoading => (
+  setLoading = isLoading =>
     this.setState({
       isLoading,
-    })
-  )
+    });
 
   /**
    * Toggles the login modal
    * Required by: <LoginModal /> <Header />
    */
   handleToggleLoginModal = () => {
-    this.setState((prevState) => ({
+    this.setState(prevState => ({
       showLoginModal: !prevState.showLoginModal,
       error: {},
     }));
-  }
+  };
 
   /**
    * handles login
    * Required by: <Header />
    */
   handleDoLogin = async (usernameValue, passwordValue) => {
-    const { username, token, error } = await makeLogin(
-      usernameValue,
-      passwordValue
-    );
+    const { username, token, error } = await makeLogin(usernameValue, passwordValue);
 
     if (username && token) {
       this.setLoggedUser(username, token);
@@ -117,7 +131,7 @@ export default class App extends Component {
         error,
       });
     }
-  }
+  };
 
   setLoggedUser = (username, token) => {
     this.setState({
@@ -125,10 +139,11 @@ export default class App extends Component {
         username,
         token,
       },
-      isUserLoggedIn: true,  // close login modal after successful login
-      showLoginModal: false,  // set isUserLoggedIn to true
+      isUserLoggedIn: true, // close login modal after successful login
+      showLoginModal: false, // set isUserLoggedIn to true
     });
-  }
+  };
+
   /**
    * Logouts user
    * Required by: <Header />
@@ -140,27 +155,7 @@ export default class App extends Component {
       user: {},
       isUserLoggedIn: false,
     });
-  }
-
-  render() {
-    const { isLoading, isUserLoggedIn, packages } = this.state;
-    return (
-      <Container isLoading={isLoading}>
-        {isLoading ? (
-          <Loading />
-        ) : (
-          <Fragment>
-            {this.renderHeader()}
-            <Content>
-              <Route isUserLoggedIn={isUserLoggedIn} packages={packages} />
-            </Content>
-            <Footer />
-          </Fragment>
-        )}
-        {this.renderLoginModal()}
-      </Container>
-    );
-  }
+  };
 
   renderLoginModal = () => {
     const { error, showLoginModal } = this.state;
@@ -173,18 +168,24 @@ export default class App extends Component {
         visibility={showLoginModal}
       />
     );
-  }
+  };
+
+  renderContent = () => {
+    return (
+      <Fragment>
+        <Content>
+          <RouterApp onLogout={this.handleLogout} onToggleLoginModal={this.handleToggleLoginModal}>
+            {this.renderHeader()}
+          </RouterApp>
+        </Content>
+        <Footer />
+      </Fragment>
+    );
+  };
 
   renderHeader = () => {
-    const { logoUrl, user, scope } = this.state;
-    return (
-      <Header
-        logo={logoUrl}
-        onLogout={this.handleLogout}
-        onToggleLoginModal={this.handleToggleLoginModal}
-        scope={scope}
-        username={user.username}
-      />
-    );
-  }
+    const { logoUrl, user: { username } = {}, scope } = this.state;
+
+    return <Header logo={logoUrl} onLogout={this.handleLogout} onToggleLoginModal={this.handleToggleLoginModal} scope={scope} username={username} />;
+  };
 }
