@@ -14,7 +14,7 @@ import { ClipBoardCopy, ClipBoardCopyText, CopyIcon } from './styles';
 import { copyToClipBoardUtility } from '../../utils/cli-utils';
 import { TEXT } from '../../utils/constants';
 
-const CopyToClipBoard = ({ text }: IProps): Node => {
+const CopyToClipBoard = ({ text, children }: IProps): Node => {
   const renderToolTipFileCopy = () => (
     <Tooltip disableFocusListener={true} title={TEXT.CLIPBOARD_COPY}>
       <CopyIcon onClick={copyToClipBoardUtility(text)}>
@@ -22,9 +22,17 @@ const CopyToClipBoard = ({ text }: IProps): Node => {
       </CopyIcon>
     </Tooltip>
   );
+
+  const renderText = children => {
+    if (children) {
+      return <ClipBoardCopyText>{children}</ClipBoardCopyText>;
+    }
+
+    return <ClipBoardCopyText>{text}</ClipBoardCopyText>;
+  };
   return (
     <ClipBoardCopy>
-      <ClipBoardCopyText>{text}</ClipBoardCopyText>
+      {renderText(children)}
       {renderToolTipFileCopy()}
     </ClipBoardCopy>
   );
