@@ -11,10 +11,10 @@ export const Wrapper = styled.span`
     vertical-align: middle;
     line-height: 22px;
     border-radius: 2px;
-    color: #9f9f9f;
-    background-color: hsla(0, 0%, 51%, 0.1);
+    color: #485a3e;
+    background-color: #f3f4f2;
     padding: 0.22rem 0.4rem;
-    margin: 5px 10px 0 0;
+    margin: 8px 8px 0 0;
     ${ellipsis('300px')};
   }
 `;
