@@ -18,7 +18,7 @@ describe('basic system test', () => {
 
   beforeAll(async function(done) {
 
-    app.use(await endPointAPI(config));
+    app.use(await endPointAPI(config()));
 
     server.listen(0, function() {
       port = server.address().port;
@@ -40,9 +40,9 @@ describe('basic system test', () => {
     });
   });
 
-  test('server should respond on /whatever', done => {
+  test('server should respond on /___not_found_package', done => {
     request({
-      url: `http://localhost:${port}/whatever`,
+      url: `http://localhost:${port}/___not_found_package`,
     }, function(err, res, body) {
       expect(err).toBeNull();
       expect(body).toMatch(API_ERROR.NO_PACKAGE);
